Add files via upload
In this project, we've learned how to set a loader properly, how to request and fetch data from Unsplash API, and how to set this data, in this case images, in HTML from JS, creating new Elements and setting the imgs properties to this elements by setAttribute. Also, we' ve learned how to create a infinite scrolling, loading new data when we get near to the end of the document.body height.